Add files via upload
Write Java Script code using HTML to resolve the given problem statmeent.
Problem : Statement
a. When we click on the openform button the Login form should be Popup with given
controls
b. When we login button then required field information is needed “ Please fill out this
filed”
c. When we click on login button , form information should be display on the Iframe with
control information.
d. When we click on the close button , popup form should be disappar.